Move function rustify_pointers to analysis
