+++ src/analysis/mod.rs
@@ -1,3 +1,4 @@
+pub mod c_type;
 pub mod functions;
 pub mod general;
 pub mod implements;
